+++ core/src/main/java/com/bumble/appyx/core/integration/NodeHost.kt
@@ -50,7 +50,10 @@ fun <N : Node> NodeHost(
             node.updateLifecycleState(source.lifecycle.currentState)
         }
         lifecycle.addObserver(observer)
-        onDispose { lifecycle.removeObserver(observer) }
+        onDispose {
+            lifecycle.removeObserver(observer)
+            integrationPoint.detach()
+        }
     }
 }