+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Pending changes
 
  - Migrate [app-tree-utils](https://github.com/badoo/app-tree-utils) into this repository.
+ - Fix integration point attached twice crash 
 
 ## 1.0-alpha01